Extract ConsoleOutputExport from ConsoleToolbar
Moves the NSSavePanel + NSPasteboard calls out of the toolbar view into
a small view-layer namespace (Brew/Features/Console/Views/ConsoleOutputExport.swift).
ConsoleToolbar's import set shrinks to SwiftUI alone, and the save/copy
buttons just call ConsoleOutputExport.save(job) / .copy(job) with data
that already lives on CommandJob.

This keeps AppKit out of the view model layer (the call sites are still
in views) while concentrating the AppKit dependency in one file rather
than scattered across toolbar button handlers.
